moved proxy tests to separate file
also fixed comment format for consistency
